Скрытие кнопок в meny.py (Не работает)
метод update у pg.Sprite не позволяет возвращать аргументы.
Можно, конечно, попробовать костыль с текстовым файлом...
Но это гемор.

Либо через глобальную переменную в самом menu.py...
