+++ utility/menu.py
@@ -1,8 +1,39 @@
 """
 Здесь будут содержаться коды ко всем меню (см. Milestone Механики)
 """
+import configparser
+
 import pygame
+import os
+import pygame as pg
 from utility.field import Board
+
+"""
+Как возвращать аргументы через функцию update у sprite ?
+"""
+def exit(*args):
+    return -1
+
+def main_menu(*args):
+    return 0
+
+def start(*args):
+    return 1
+
+
+def load_image(name, resize_ch=1):
+    fullname = os.path.join('textures', name)
+    try:
+        image = pg.image.load(fullname)
+    except pg.error as message:
+        print('Cannot load image:', name)
+        raise SystemExit(message)
+
+    if resize_ch != 1:
+        img_size = image.get_size()
+        new_img = (img_size[0] * resize_ch, img_size[1] * resize_ch)
+        image = pg.transform.scale(image, new_img)
+    return image
 
 
 class Button(pygame.sprite.Sprite):
@@ -36,6 +67,8 @@ def start_screen(screen: pygame.Surface, len_side_screen: int, count_cells: int)
     """
     background_render(screen, len_side_screen, count_cells)
 
+    start_screen_buttons.draw(screen)
+
     intro_text = ["Snake 2.0"]
     font = pygame.font.Font(None, 100)
     text_coord = 50
@@ -59,7 +92,24 @@ def level_screen(screen: pygame.Surface, len_side_screen: int, count_cells: int)
 
     """
     background_render(screen, len_side_screen, count_cells)
+    level_screen_buttons.draw(screen)
 
 
 def gameover_screen(screen: pygame.Surface, len_side_screen: int, count_cells: int, len_snake: int) -> None:
     pass
+
+
+config = configparser.ConfigParser()
+config.read('config.cfg')
+len_side_screen = int(config['screen']['len_side'])
+count_cells = int(config['screen']['count_cells'])
+
+start_screen_buttons = pg.sprite.Group()
+level_screen_buttons = pg.sprite.Group()
+gameover_screen_buttons = pg.sprite.Group()
+exit_button = Button(len_side_screen // 2, len_side_screen // 2 + ((len_side_screen // 2) // 3) * 2,
+                     load_image("end.png", 7), exit, start_screen_buttons)
+start_button = Button(len_side_screen // 2, len_side_screen // 2, load_image("start.png", 7), start,
+                      start_screen_buttons)
+
+list_groups_buttons = [start_screen_buttons, level_screen_buttons, gameover_screen_buttons]
